+++ packages/assets/src/wallets/wallet-connect.tsx
@@ -1,9 +1,9 @@
 // 说明：WalletConnect 通用协议的元数据展示（支持移动端扫码连接）
 import type { WalletMetadata } from 'pelican-web3-lib-common';
-import { WalletConnectColorful } from 'pelican-web3-lib-icons';
+import { WalletConnectArk } from 'pelican-web3-lib-icons';
 
 export const metadata_WalletConnect: WalletMetadata = {
-  icon: <WalletConnectColorful />,
+  icon: <WalletConnectArk />,
   name: 'WalletConnect',
   remark: 'Connect with mobile APP',
   universalProtocol: {