+++ packages/icons/src/components/metamask-ark.tsx
@@ -1,0 +1,26 @@
+// GENERATE BY ./scripts/generate.ts
+// DON NOT EDIT IT MANUALLY
+import * as React from 'react';
+import AntdIcon from '@ant-design/icons';
+import { type IconBaseProps } from '@ant-design/icons/lib/components/Icon';
+import { ConfigProvider } from 'antd';
+import classnames from 'classnames';
+
+import { eraseDefaultProps } from '../utils';
+import SVGComponent from '../svgs/ark-metamask.svg';
+
+export const MetaMaskArk = React.forwardRef<HTMLSpanElement, IconBaseProps>((props, ref) => {
+  const { getPrefixCls } = React.useContext(ConfigProvider.ConfigContext);
+  const prefixCls = getPrefixCls('web3-icon-metamask-colorful');
+
+  return (
+    <AntdIcon
+      {...props}
+      className={classnames(prefixCls, props.className)}
+      ref={ref}
+      component={eraseDefaultProps(SVGComponent)}
+    />
+  );
+});
+
+MetaMaskArk.displayName = 'MetaMaskArk';