+++ examples/eth-web3js/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @example/eth-web3js
 
+## 0.0.29
+
+### Patch Changes
+
+- Updated dependencies
+  - pelican-web3-lib-assets@1.0.3
+  - pelican-web3-lib-common@1.0.3
+  - pelican-web3-lib-eth-web3js@1.0.3
+  - pelican-web3-lib@1.26.3
+
 ## 0.0.28
 
 ### Patch Changes