+++ packages/ethers/CHANGELOG.md
@@ -1,5 +1,15 @@
 # pelican-web3-lib-ethers
 
+## 1.0.3
+
+### Patch Changes
+
+- 批量发布：对所有非 web3 的包进行补丁版本提升到 1.0.3。
+- Updated dependencies
+  - pelican-web3-lib-assets@1.0.3
+  - pelican-web3-lib-common@1.0.3
+  - pelican-web3-lib-evm@1.0.3
+
 ## 1.0.2
 
 ### Patch Changes