+++ packages/web3/CHANGELOG.md
@@ -1,5 +1,14 @@
 # pelican-web3-lib
 
+## 1.26.3
+
+### Patch Changes
+
+- Updated dependencies
+  - pelican-web3-lib-assets@1.0.3
+  - pelican-web3-lib-common@1.0.3
+  - pelican-web3-lib-icons@1.0.3
+
 ## 1.26.2
 
 ### Patch Changes